% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{agglomerative_clustering}
\alias{agglomerative_clustering}
\title{Agglomerative clustering}
\usage{
agglomerative_clustering(df, clust_melt_rad, iterations, cores = 2)
}
\arguments{
\item{df}{A tibble containing 3D point coordinates of vertices that
#' lie above threshold. Should contain the columns \verb{x, y, z}.
#' Typically, this is the resulting tibble of the \code{get_local_height()} function.}

\item{clust_melt_rad}{A numerical value of search radius for agglomerative
clustering.}

\item{iterations}{A numerical value of how many clustering iterations should
be run. Default: \code{1}.}

\item{cores}{Number of CPU cores to use. Default = \code{2}.}
}
\description{
xxx: add description
}
\examples{
xxx: add example

}
