% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{angle_between_vectors}
\alias{angle_between_vectors}
\title{Calculate angle between two vectors in 3D}
\usage{
angle_between_vectors(a, b)
}
\arguments{
\item{a}{A \code{vector} with three \code{numeric} data points describing vector 1.}

\item{b}{A \code{vector} with three \code{numeric} data points describing vector b.}
}
\value{
Returns a \code{numeric} value with the angle in degree (°).
}
\description{
xxx: add description
}
\examples{
xxx: add example

}
