% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{average_coordinates}
\alias{average_coordinates}
\title{Average x, y, and z coordinates of facets according to their neighbours}
\usage{
average_coordinates(df, verbose = FALSE)
}
\arguments{
\item{df}{A tibble containing facet coordinates in columns \verb{x, y, z}
as well as neighbours in a column \code{neighbours}.}

\item{verbose}{A \code{logical} value indicating if message printing is permitted.
Default: \code{FALSE}.}
}
\value{
Returns a \code{tibble} containing the additional columns with info on
the Eye Parameter (P), the inter-facet angle (delta.phi) and acuity (CPD) for
each facet.
}
\description{
xxx: add description
}
\examples{
xxx: add example

}
