% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_delta.phi}
\alias{calc_delta.phi}
\title{Calculates the angle between 2 vectors}
\usage{
calc_delta.phi(curr.lens, curr.partner, type)
}
\arguments{
\item{df}{A tibble containing coordinates in columns \verb{x, y, z}.}

\item{line_points}{A 2x3 tibble containing coordinates of line to align the
point cloud to. Must contain one row per point and columns \verb{x, y, z}.}

\item{axis}{A character string defining the global axis to align to. Must be
\code{x}, \code{y}, or \code{z}.}
}
\value{
Returns a tibble with the aligned coordinates in columns \verb{x, y, z}.
}
\description{
Rotates 3D point cloud according to one defined vector so that this vector
is aligned to one of the global coordinate system axes.
}
\examples{
# xxx: add example and change above descsriptionand parameters

}
