% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfklow.R
\name{calculate_corneal_projections}
\alias{calculate_corneal_projections}
\title{Calculate Corneal Projections}
\usage{
calculate_corneal_projections(df, cp_diam_cm, center = NULL, verbose = FALSE)
}
\arguments{
\item{df}{A tibble containing coordinates and IDs.}
}
\value{
Returns a tibble with the aligned coordinates in columns defined by
\verb{x_col, y_col, z_col}.
}
\description{
XYZ: create description and param defs
}
\examples{
xxx: add example and change above descsriptionand parameters

}
