% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_heights.R
\name{combine_facets_and_LMs}
\alias{combine_facets_and_LMs}
\title{Combine facets with landmarks}
\usage{
combine_facets_and_LMs(
  df,
  local_heights,
  landmarks_file,
  crop_log_file_name,
  cores,
  plot_results = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{A tibble containing triangle center coordinates in columns \verb{x, y, z}.}

\item{cores}{A numerical value of how many cores to use. Default: \code{1}.}

\item{search_diam}{A numerical value defining the size of the search diameter
of defining the local plane.}
}
\value{
Tibble \code{df} with additional column \code{local_height}.
}
\description{
Calculate distance of vertices from local plane.
}
\examples{
# xxx: add example

}
