% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfklow.R
\name{convert_STL_to_tibble}
\alias{convert_STL_to_tibble}
\title{Convert STL to tibble}
\usage{
convert_STL_to_tibble(
  file_name,
  stl_folder,
  triangle_centers_and_normals_folder,
  plot_results = TRUE,
  write_results_to_file = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{df}{A tibble containing coordinates and IDs.}
}
\value{
Returns a tibble with the aligned coordinates in columns defined by
\verb{x_col, y_col, z_col}.
}
\description{
XYZ: create description and param defs
}
\examples{
xxx: add example and change above descsriptionand parameters

}
