% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_log_data}
\alias{crop_log_data}
\title{Data of cropping log file}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{ID}{A character vector of variable names}
\item{x}{A character vector of variable values}
}
}
\source{
Generated using \code{crop_scan.ijm} script of compoundvision3D.
}
\usage{
crop_log_data
}
\description{
This is an example tibble of a \code{crop.log} file.
}
\keyword{datasets}
