% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cv3d_example_facets}
\alias{cv3d_example_facets}
\title{Example Checked Facet Centres}
\format{
A data frame with 61 rows and 4 variables:
\describe{
\item{ID}{Sequential facet identifier.}
\item{x, y, z}{Three-dimensional facet-centre coordinates in micrometres.}
}
}
\source{
Schoborg TA, Smith SL, Smith LN, Morris HD, Rusan NM (2019).
Micro-computed tomography as a platform for exploring Drosophila
development. \emph{Development} 146(23): dev176685.
\doi{10.1242/dev.176685}
}
\usage{
cv3d_example_facets
}
\description{
A reduced contiguous subset of checked compound-eye facet centres from an
adult \emph{Drosophila melanogaster}. The source micro-computed tomography data
are described by Schoborg et al. (2019). The dataset is intended for
examples of facet-neighbour, facet-size, facet-normal, and optical-property
calculations.
}
\keyword{datasets}
