% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cv3d_example_surface}
\alias{cv3d_example_surface}
\title{Example Compound-Eye Surface Mesh}
\format{
A data frame with 600 rows and 7 variables:
\describe{
\item{ID}{Sequential triangle identifier.}
\item{x, y, z}{Three-dimensional triangle-centre coordinates in micrometres.}
\item{norm.x, norm.y, norm.z}{Components of the triangle surface normal.}
}
}
\source{
Schoborg TA, Smith SL, Smith LN, Morris HD, Rusan NM (2019).
Micro-computed tomography as a platform for exploring Drosophila
development. \emph{Development} 146(23): dev176685.
\doi{10.1242/dev.176685}
}
\usage{
cv3d_example_surface
}
\description{
A reduced contiguous subset of a compound-eye surface mesh from an adult
\emph{Drosophila melanogaster}. The source micro-computed tomography data are
described by Schoborg et al. (2019). The dataset contains triangle centres
and triangle surface normals and is intended for examples involving local
surface-height calculations.
}
\keyword{datasets}
