% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cv3d_example_thresholded}
\alias{cv3d_example_thresholded}
\title{Example Thresholded Local-Height Points}
\format{
A data frame with 74 rows and 5 variables:
\describe{
\item{source_index}{Sequential point identifier.}
\item{x, y, z}{Three-dimensional point coordinates in micrometres.}
\item{height_value}{Thresholded local surface-height value.}
}
}
\source{
Schoborg TA, Smith SL, Smith LN, Morris HD, Rusan NM (2019).
Micro-computed tomography as a platform for exploring Drosophila
development. \emph{Development} 146(23): dev176685.
\doi{10.1242/dev.176685}
}
\usage{
cv3d_example_thresholded
}
\description{
A reduced subset of thresholded local-height points from the same adult
\emph{Drosophila melanogaster} compound eye represented by
\code{cv3d_example_surface}. The source micro-computed tomography data are
described by Schoborg et al. (2019). The dataset is intended for examples of
facet-candidate condensation.
}
\keyword{datasets}
