% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{distance_3D}
\alias{distance_3D}
\title{Calculate distance between two points in 3D.}
\usage{
distance_3D(point1, point2, verbose = FALSE)
}
\arguments{
\item{point1}{A \code{vector} containing the \code{numeric} \verb{x, y} and \code{z} coordinates
pf point 1.}

\item{point2}{A \code{vector} containing the \code{numeric} \verb{x, y} and \code{z} coordinates
pf point 2.
#' @param verbose A \code{logical} value indicating if message printing is permitted.
Default: \code{FALSE}.}
}
\value{
Returns the \code{numeric} distance between point 1 and point 2.
}
\description{
xxx: add description
}
\examples{
xxx: add example

}
