% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facet_positions}
\alias{facet_positions}
\title{Facet Positions}
\format{
A data frame with 735 rows and 4 variables:
\describe{
\item{ID}{A numeric vector of consequtive facet IDs}
\item{x}{A numeric vector of facet x coordinates}
\item{y}{A numeric vector of facet y coordinates}
\item{z}{A numeric vector of facet z coordinates}
}
}
\source{
Generated using the software Blender and the Blender Python
script \verb{4_export_corrected_facets.py}.
}
\usage{
facet_positions
}
\description{
This is an example tibble containing 735 facet coordinates.
}
\keyword{datasets}
