% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facet_positions_new}
\alias{facet_positions_new}
\title{Facet Positions Referenced}
\format{
A data frame with 735 rows and 8 variables:
\describe{
\item{ID}{A numeric vector of consequtive facet IDs}
\item{x}{A numeric vector of facet x coordinates}
\item{y}{A numeric vector of facet y coordinates}
\item{z}{A numeric vector of facet z coordinates}
\item{norm.x}{A numeric vector of vertex  facet x components}
\item{norm.y}{A numeric vector of vertex  facet y components}
\item{norm.z}{A numeric vector of vertex  facet z components}
\item{local_height}{A numeric vector of the facet local height}
}
}
\source{
Generated using \code{compoundvision3D::reference_point_clouds()}.
}
\usage{
facet_positions_new
}
\description{
This is an example tibble with 735 facet coordinates that where already
referenced to the \code{local_heights} tibble.
}
\keyword{datasets}
