% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_diam_definition.R
\name{facet_size_diam_interactive}
\alias{facet_size_diam_interactive}
\title{Search Diameter Definition}
\usage{
facet_size_diam_interactive(df)
}
\arguments{
\item{df}{A tibble containing triangle center coordinates in columns \verb{x, y, z}.}
}
\value{
A numerical search diameter.
}
\description{
Define local plane search diameter interactively.
}
\examples{
# xxx: add example

}
