% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{find_facet_canidates}
\alias{find_facet_canidates}
\title{Fine clustering towards Facet Peaks}
\usage{
find_facet_canidates(...)
}
\description{
Backwards-compatible wrapper for the historical misspelled function name.
}
