% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{find_facets_fine}
\alias{find_facets_fine}
\title{Fine clustering towards Facet Peaks}
\usage{
find_facets_fine(
  df,
  cols_to_use,
  h_min = NULL,
  h_max = NULL,
  h_final = NULL,
  n_steps = 100,
  plot_file = NULL
)
}
\arguments{
\item{df}{A tibble containing triangle center coordinates of vertices that
lie above threshold in columns \verb{x, y, z} .
Typically, this is the resulting tibble of the \code{find_facets_rough()}
function.}

\item{cols_to_use}{A numerical value of which columns store x,y,z coordinates
of rough clustering.}
}
\value{
Tibble \code{df} with ideally one vertex per facet, located on the facet
peak.
}
\description{
Find facet peaks.
}
\examples{
xxx: add example

}
