% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfklow.R
\name{find_local_heights}
\alias{find_local_heights}
\title{Find local heights}
\usage{
find_local_heights(
  file_name,
  triangle_centers_and_normals_folder,
  local_heights_folder,
  local_heights_folder_normalized,
  curr_facet_estimate = 14,
  cores,
  plot_results = TRUE,
  write_results_to_file = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{df}{A tibble containing coordinates and IDs.}
}
\value{
Returns a tibble with the aligned coordinates in columns defined by
\verb{x_col, y_col, z_col}.
}
\description{
XYZ: create description and param defs
}
\examples{
xxx: add example and change above descsriptionand parameters

}
