% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{find_threshold}
\alias{find_threshold}
\title{Find threshold for agglomerative clustering}
\usage{
find_threshold(
  df = local_heights,
  column1 = NULL,
  column2 = NULL,
  height_column,
  min_threshold = 1,
  max_treshold = 2.5,
  trials = 12,
  plot_file = NULL
)
}
\description{
Very roughly find clusters around facet peaks.
}
\details{
find good threshold - choose x,y, x,z or other dimension combinations to get reasonable plots
}
