% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_height.R
\name{get_local_heights}
\alias{get_local_heights}
\title{Find Local Vertex Heights}
\usage{
get_local_heights(
  df,
  search_diam,
  cores = 1,
  log_scale = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{A tibble containing triangle center coordinates in columns \verb{x, y, z}.}

\item{search_diam}{A numerical value defining the size of the search diameter
of defining the local plane.}

\item{cores}{A numerical value of how many cores to use. Default: \code{1}.}
}
\value{
Tibble \code{df} with additional column \code{local_height}.
}
\description{
Calculate distance of vertices from local plane.
}
\examples{
# xxx: add example

}
