% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfklow.R
\name{get_optic_parameters}
\alias{get_optic_parameters}
\title{Calculate optic parameters for eye}
\usage{
get_optic_parameters(
  df,
  plot_file = NULL,
  facet_estimate = 14,
  cores = 18,
  plot_results = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{A tibble containing coordinates and IDs.}
}
\value{
Returns a tibble with the aligned coordinates in columns defined by
\verb{x_col, y_col, z_col}.
}
\description{
XYZ: create description and param defs
}
\examples{
xxx: add example and change above descsriptionand parameters

}
