% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_plot_axes}
\alias{get_plot_axes}
\title{Get optimal global viewing axis}
\usage{
get_plot_axes(df)
}
\arguments{
\item{df}{Dataframe with facet coordinates.}
}
\value{
xxx
}
\description{
xxx
}
\examples{
xxx: add example and change above parameter description

}
