% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LMs_df}
\alias{LMs_df}
\title{Landmarks}
\format{
A data frame with 9 rows and 4 variables:
\describe{
\item{LM}{A character vector of landmark names}
\item{x}{A numeric vector of landmark x coordinates}
\item{y}{A numeric vector of landmark y coordinates}
\item{z}{A numeric vector of landmark z coordinates}
}
}
\source{
Generated using the software 3D SLicer.
}
\usage{
LMs_df
}
\description{
This is an example tibble containing landmark coordinates.
}
\keyword{datasets}
