% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{local_heights}
\alias{local_heights}
\title{Local Heights}
\format{
A data frame with 22,464 rows and 8 variables:
\describe{
\item{ID}{A numeric vector of consequtive vertex IDs}
\item{x}{A numeric vector of vertex x coordinates}
\item{y}{A numeric vector of vertex y coordinates}
\item{z}{A numeric vector of vertex z coordinates}
\item{norm.x}{A numeric vector of vertex  normal x components}
\item{norm.y}{A numeric vector of vertex  normal y components}
\item{norm.z}{A numeric vector of vertex  normal z components}
\item{local_height}{A numeric vector of the vertex local height}
}
}
\source{
Generated using \code{compoundvision3D::get_local_heights()} on
\code{compoundvision3D::Drosohila_eye_tri_centers_normals}
}
\usage{
local_heights
}
\description{
This is an example a tibble containing information on local heights of each
vertex.
}
\keyword{datasets}
