% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{move_point_along_vector}
\alias{move_point_along_vector}
\title{Move a 3D point along a vector}
\usage{
move_point_along_vector(point, vector, distance)
}
\arguments{
\item{point}{A \code{tibble} or \code{vector} containing the \verb{x, y, z} coordinates of the point to move.}

\item{vector}{A \code{tibble} or \code{vector} containing the \verb{x, y, z} values of the vector.}

\item{distance}{A \code{numeric} value by which the point shall be moved.}
}
\value{
Returns a \code{vector} containing the \verb{x, y, z} coordinates of the moved point.
}
\description{
xxx: add description
}
\examples{
xxx: add example
}
