% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfklow.R
\name{normalize_eye_data}
\alias{normalize_eye_data}
\title{Normalize data of all eyes}
\usage{
normalize_eye_data(df, info_table = CompVisTab, plot_results = FALSE)
}
\arguments{
\item{df}{A tibble containing coordinates and IDs.}
}
\value{
Returns a tibble with the aligned coordinates in columns defined by
\verb{x_col, y_col, z_col}.
}
\description{
XYZ: create description and param defs
}
\examples{
xxx: add example and change above descsriptionand parameters

}
