% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_eye}
\alias{plot_eye}
\title{Plot eye in 3D}
\usage{
plot_eye(
  facets_coords,
  facet_colours = "red",
  facet_sizes = 10,
  facet_type = "p",
  LM_coords = NULL,
  LM_colours = "blue",
  LM_sizes = 10,
  LM_type = "p",
  text_coords = NULL,
  text_labels = NULL
)
}
\arguments{
\item{file_name}{File name of STL to import.}
}
\value{
A tibble containing triangle centers and triangle normals of STL file.
}
\description{
Imports triangle centers and triangle normals of STL file as tibble. We use
the word 'triangle' here to refer to the facets of an STL mesh to avoid
confusion with the facets of compound eyes.
}
\examples{
# xxx: add example and change above parameter description

}
