% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_facet_and_neighbours}
\alias{plot_facet_and_neighbours}
\title{Plot a facet with it's neighbours}
\usage{
plot_facet_and_neighbours(df, facet, radius = 1, threeD = FALSE)
}
\arguments{
\item{df}{Dataframe with facet numbers and their coordinates.}

\item{facet}{The number of the facet.}

\item{radius}{Radius to plot the spheres.}

\item{threeD}{\code{Boolean} value for also plotting in 3D or just in 2D. Default: \code{FALSE} (= just 2D).}
}
\value{
xxx
}
\description{
xxx
}
\examples{
xxx: add example and change above parameter description

}
