% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read_3DSlicer_landmarks}
\alias{read_3DSlicer_landmarks}
\title{Read json File with Landmark Data from 3D Slicer}
\usage{
read_3DSlicer_landmarks(file)
}
\arguments{
\item{file}{A string containing the full path of the json file.}
}
\value{
Returns a tibble with the landmark data in columns
\verb{LM,x, y, z}.
}
\description{
Reads a *.mrk.json file containing landmarks created in 3D slicer.
}
\examples{
# xxx: add example

}
