% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_plus}
\alias{read_plus}
\title{Read multiple csv files and store their file name in column}
\usage{
read_plus(file_list, show_col_types = FALSE)
}
\arguments{
\item{file_list}{A vector containing \code{character} strings of csv file names.}

\item{show_col_types}{A \code{logic} value to define if column types should be
shown while importing. \code{Default: FALSE}.}

\item{axis}{A character string defining the global axis to align to. Must be
\code{x}, \code{y}, or \code{z}.}
}
\value{
Returns a tibble with the columns as they appear in all csv files.
}
\description{
Read multiple csv files and store their file name in column
}
\examples{
xxx: add example and change above descsriptionand parameters

}
