% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_facets.R
\name{reference_point_clouds}
\alias{reference_point_clouds}
\title{References Coordinates to Exisiting Point Cloud.}
\usage{
reference_point_clouds(
  coordinates,
  reference_coordinates,
  search_diameter = 10,
  cores = 1
)
}
\arguments{
\item{coordinates}{A dataframe containing at least the columns \verb{x,y,z}.}

\item{reference_coordinates}{A dataframe containing at least the columns \verb{x,y,z}.}

\item{search_diameter}{A numeric value to define the size of the box that is
used around each coordinate to find its closest reference coordinate.
Defailt: \code{1}.}
}
\value{
Returns a tibble with points in \code{reference_coordinates} that are
closest to the points in \code{coordinates}, containing the same columns as
\code{reference_coordinates}.
}
\description{
Takes a point cloud of coordinates, e.g. facet positions, and finds the
closest points (including their data) in a second point cloud of coordinates.
}
\examples{
# xxx: add example

}
