% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfklow.R
\name{rotate_and_translate_to_global}
\alias{rotate_and_translate_to_global}
\title{Rotate and translate facet positions according to the global coordinate system}
\usage{
rotate_and_translate_to_global(
  df,
  facet_estimate = 14,
  cores = 18,
  plot_results = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{df}{A tibble containing coordinates and IDs.}
}
\value{
Returns a tibble with the aligned coordinates in columns defined by
\verb{x_col, y_col, z_col}.
}
\description{
XYZ: create description and param defs
}
\examples{
xxx: add example and change above descsriptionand parameters

}
