% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rotate_rgl_to_vector}
\alias{rotate_rgl_to_vector}
\title{Rotate the rgl viewport to face a specific direction vector}
\usage{
rotate_rgl_to_vector(direction_vector)
}
\arguments{
\item{direction_vector}{Vector with three numbers indicating the desired
viewing direction.}
}
\value{
xxx
}
\description{
xxx
}
\examples{
xxx: add example and change above parameter description

}
