% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{set_up_folder_structure}
\alias{set_up_folder_structure}
\title{Create folders}
\usage{
set_up_folder_structure(parent_folder = NULL, verbose = TRUE)
}
\arguments{
\item{parent_folder}{Path to where folders should be created.
If \code{NULL}, current working directory will be used.
Default: \code{NULL}.}
}
\value{
Creates the following folders:\cr
\verb{./data/}\cr
\verb{./data/1_pre_STLs/}\cr
\verb{./data/2_STLs/}\cr
\verb{./data/3_triangle_centers_and_normals/}\cr
\verb{./data/4_1_local_heights/}\cr
\verb{./data/4_2_local_heights_normalized/}\cr
\verb{./data/5_rough_clusters/}\cr
\verb{./data/6_facet_candidates/}\cr
\verb{./data/7_facet_positions/}\cr
\verb{./data/8_facets_landmarks_combined/}\cr
\verb{./data/9_facets_landmarks_rotated/}\cr
\verb{./data/10_facet_infos/}\cr
\verb{./data/11_analyses/}
}
\description{
Creates folders in folder structure expected by the
CompoundVision3D package.
}
\examples{
xxx: add example

}
