% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_to_R.R
\name{STL_vertices}
\alias{STL_vertices}
\title{Import STL vertices as Tibble}
\usage{
STL_vertices(file_name)
}
\arguments{
\item{file_name}{File name of STL to import.}
}
\value{
A tibble containing vertex coordinates of STL file.
}
\description{
Imports vertex coordinates of STL file as tibble.
}
\examples{
xxx: add example


}
