% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfklow.R
\name{threshold_high_points}
\alias{threshold_high_points}
\title{Filter high cuticle vertices by thresholding}
\usage{
threshold_high_points(
  df,
  thresholds_file,
  column1 = "x",
  column2 = "y",
  min_threshold = 2,
  max_treshold = 8,
  final_threshold = 5,
  trials = 9,
  cores = 18,
  plot_results = TRUE,
  plot_file,
  verbose = FALSE
)
}
\arguments{
\item{df}{A tibble containing coordinates and IDs.}
}
\value{
Returns a tibble with the aligned coordinates in columns defined by
\verb{x_col, y_col, z_col}.
}
\description{
XYZ: create description and param defs
}
\examples{
xxx: add example and change above descsriptionand parameters

}
