% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tri_centers_normals}
\alias{tri_centers_normals}
\title{Triangle Centers and Normals}
\format{
A data frame with 22,464 rows and 7 variables:
\describe{
\item{ID}{A numeric vector of consequtive vertex IDs}
\item{x}{A numeric vector of vertex x coordinates}
\item{y}{A numeric vector of vertex y coordinates}
\item{z}{A numeric vector of vertex z coordinates}
\item{norm.x}{A numeric vector of vertex  normal x components}
\item{norm.y}{A numeric vector of vertex  normal y components}
\item{norm.z}{A numeric vector of vertex  normal z components}
}
}
\source{
Generated using \code{compoundvision3D::STL_triangles()}
}
\usage{
tri_centers_normals
}
\description{
This is an example an imported STL file.
}
\keyword{datasets}
