% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{align_pointcloud}
\alias{align_pointcloud}
\title{Align 3D Point Cloud to Global Axes}
\usage{
align_pointcloud(
  df,
  ref_x = "x",
  ref_y = "y",
  ref_z = "z",
  data_x = "norm.x",
  data_y = "norm.y",
  data_z = "norm.z",
  landmark_col = "ID",
  names = list(anterior = "anterior", posterior = "posterior", left = "left", right =
    "right"),
  priority = "RL"
)
}
\arguments{
\item{df}{A tibble containing coordinates and IDs.}

\item{ref_x}{The column containing x-coordinates to be rotated. Default: \code{x}.}

\item{ref_y}{The column containing y-coordinates to be rotated. Default: \code{y}.}

\item{ref_z}{The column containing z-coordinates to be rotated. Default: \code{z}.}

\item{data_x}{The column containing the reference x-coordinates. Default: \code{x}.}

\item{data_y}{The column containing the reference y-coordinates. Default: \code{y}.}

\item{data_z}{The column containing the reference z-coordinates. Default: \code{z}.}

\item{landmark_col}{The column that has to have cells following the
vector defined in \code{names}. Default: \code{ID}.}

\item{names}{A character string defining the global axis to align to.
. Default: \verb{anterior, posterior, left, right}.}

\item{priority}{A character string defining the global axis that has priority
when aligning. Can read \code{RL} for the right-left vector or \code{AP} for the
antrior-posterior vector. Default: \code{RL}.}
}
\value{
Returns a tibble with the aligned coordinates in columns defined by
\verb{x_col, y_col, z_col}.
}
\description{
Rotates 3D point cloud according to two defined vectors so that these vectors
are aligned to the global coordinate system X- and Y-axes.
}
\examples{
xxx: add example and change above descsriptionand parameters

}
