% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_facet_size}
\alias{calculate_facet_size}
\title{Calculate facet sizes according to their neighbours}
\usage{
calculate_facet_size(
  df,
  id_col = "ID",
  x_col = "x",
  y_col = "y",
  z_col = "z",
  neighbours_col = "neighbours",
  sep = ";",
  keep_distances = FALSE
)
}
\arguments{
\item{df}{A tibble containing facet coordinates in columns \verb{x, y, z}.}

\item{cores}{A numerical value of how many cores to use. Default: \code{1}.}

\item{verbose}{A \code{logical} value indicating if message printing is permitted.
Default: \code{FALSE}.}
}
\value{
Returns a \code{tibble} containing the additional columns with info on
facet normals in x, y, and z direction for each facet.
}
\description{
xxx: add description
}
\examples{
xxx: add example

}
