% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{calculate_normal}
\alias{calculate_normal}
\title{Calculate normal of triangle in 3D}
\usage{
calculate_normal(A, B, C, normalize = TRUE)
}
\arguments{
\item{A}{A \code{vector} with the \code{numeric} x, y, and z coordinges of point 1.}

\item{B}{A \code{vector} with the \code{numeric} x, y, and z coordinges of point 2.}

\item{C}{A \code{vector} with the \code{numeric} x, y, and z coordinges of point 3.}
}
\value{
Returns a \code{vector} with three \code{numeric} data points describing the
normal vector of the triangle.
}
\description{
xxx: add description
}
\examples{
xxx: add example

}
