% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfklow.R
\name{find_facet_positions}
\alias{find_facet_positions}
\title{Find facet poisition candidates}
\usage{
find_facet_positions(
  df,
  local_heights,
  h_min = 46.127,
  h_max = 16.287,
  h_final = 23.873,
  column1 = NULL,
  column2 = NULL,
  cutoffs_file,
  plot_results = TRUE,
  plot_file,
  verbose = FALSE
)
}
\arguments{
\item{df}{A tibble containing coordinates and IDs.}
}
\value{
Returns a tibble with the aligned coordinates in columns defined by
\verb{x_col, y_col, z_col}.
}
\description{
XYZ: create description and param defs
}
\examples{
xxx: add example and change above descsriptionand parameters

}
