% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{find_neighbours}
\alias{find_neighbours}
\title{Find facet neighbors and Calculate facet sizes}
\usage{
find_neighbours(
  df,
  facet_size,
  neighbour_threshold = 1.5,
  cores = 1,
  plot_file = NULL,
  verbose = FALSE
)
}
\arguments{
\item{df}{A \code{tibble} containing facet coordinates in columns \verb{x, y, z}.}

\item{facet_size}{A \code{numeric} value containing the estimated facet size.}

\item{cores}{A numerical value of how many cores to use. Default: \code{1}.}

\item{verbose}{A \code{logical} value indicating if message printing is permitted.
Default: \code{FALSE}.}
}
\value{
Returns a \code{tibble} containing the additional columns with info on
facet size, facet neighbours and the number of neighbors of each facet.
}
\description{
xxx: add description
}
\examples{
xxx: add example

}
