% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_facet_normals}
\alias{get_facet_normals}
\title{Calculate facet normals according to their spacial positions}
\usage{
get_facet_normals(df, cores = 1, plot_file = NULL, verbose = FALSE)
}
\arguments{
\item{df}{A tibble containing facet coordinates in columns \verb{x, y, z}.}

\item{cores}{A numerical value of how many cores to use. Default: \code{1}.}

\item{verbose}{A \code{logical} value indicating if message printing is permitted.
Default: \code{FALSE}.}
}
\value{
Returns a \code{tibble} containing the additional columns with info on
facet normals in x, y, and z direction for each facet.
}
\description{
xxx: add description
}
\examples{
xxx: add example

}
