% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_height.R
\name{get_height_colors}
\alias{get_height_colors}
\title{Get colour values !HERE: Change description}
\usage{
get_height_colors(heights, lower_quantile = NULL, upper_quantile = NULL)
}
\arguments{
\item{df}{A tibble containing triangle center coordinates in columns \verb{x, y, z}.}

\item{search_diam}{A numerical value defining the size of the search diameter
of defining the local plane.}

\item{cores}{A numerical value of how many cores to use. Default: \code{1}.}
}
\value{
Tibble \code{df} with additional column \code{local_height}.
}
\description{
Calculate distance of vertices from local plane.
}
\examples{
# xxx: add example

}
