% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_optic_properties}
\alias{get_optic_properties}
\title{Get optic parameter approximations}
\usage{
get_optic_properties(df, cores = 1, verbose = FALSE)
}
\arguments{
\item{df}{A tibble containing facet coordinates in columns \verb{x, y, z}.}

\item{cores}{A numerical value of how many cores to use. Default: \code{1}.}

\item{verbose}{A \code{logical} value indicating if message printing is permitted.
Default: \code{FALSE}.}
}
\value{
Returns a \code{tibble} containing the additional columns with info on
the Eye Parameter (P), the inter-facet angle (delta.phi) and acuity (CPD) for
each facet.
}
\description{
xxx: add description
}
\examples{
xxx: add example

}
