% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_heights.R
\name{normalize_local_heights}
\alias{normalize_local_heights}
\title{Noramlize Local Vertex Heights}
\usage{
normalize_local_heights(
  df,
  normalize_diam,
  column_to_normalize = "local_height",
  cores = 12,
  plot_file = NULL,
  verbose = FALSE
)
}
\arguments{
\item{df}{A tibble containing triangle center coordinates in columns \verb{x, y, z}.}

\item{cores}{A numerical value of how many cores to use. Default: \code{1}.}

\item{search_diam}{A numerical value defining the size of the search diameter
of defining the local plane.}
}
\value{
Tibble \code{df} with additional column \code{local_height}.
}
\description{
xxx: update! Calculate distance of vertices from local plane.
}
\examples{
# xxx: add example

}
