% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_to_R.R
\name{STL_triangles}
\alias{STL_triangles}
\title{Import STL triangles (facets) as Tibble}
\usage{
STL_triangles(file_name)
}
\arguments{
\item{file_name}{File name of STL to import.}
}
\value{
A tibble containing triangle centers and triangle normals of STL file.
}
\description{
Imports triangle centers and triangle normals of STL file as tibble. We use
the word 'triangle' here to refer to the facets of an STL mesh to avoid
confusion with the facets of compound eyes.
}
\examples{
# xxx: add example

}
